ciyunapi/requirements.txt
amqp==5.0.6
asgiref==3.3.4
beautifulsoup4==4.9.3
billiard==3.6.4.0
bs4==0.0.1
celery==5.0.5
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cryptography==3.4.7
cycler==0.10.0
Django==3.2
django-cors-headers==3.7.0
django-json-rpc==0.7.2
djangorestframework==3.12.4
h11==0.12.0
idna==2.10
imageio==2.9.0
jieba==0.42.1
kiwisolver==1.3.1
kombu==5.0.2
lxml==4.6.3
matplotlib==3.4.1
multidict==5.1.0
numpy==1.20.2
Pillow==8.2.0
prompt-toolkit==3.0.18
pycparser==2.20
pymongo==3.11.3
PyMySQL==1.0.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.26.4
utils==1.0.1
uvicorn==0.13.4
vine==5.0.0
wcwidth==0.2.5
wordcloud
